pyproject.toml
[tool.poetry]
name = "script-handler"
version = "0.0.0"
description = "Streamlit script and service handler"
authors = ["FloWide Ltd."]
readme = "README.md"
packages = [{include = "script_handler"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
pydantic = "^1.10.7"
pydantic-argparse = "^0.8.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pyjwt = {extras = ["crypto"], version = "^2.6.0"}
aiohttp = "^3.8.4"
python-lsp-server = {extras = ["pycodestyle", "pyflakes","yapf"], version = "^1.7.2"}
aiofiles = "^23.1.0"
loguru = "^0.7.0"
sortedcontainers = "^2.4.0"
pyyaml = "^6.0"
pygit2 = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry_bumpversion.file."script_handler/_version.py"]
